Moved `examples/containers` to `tests/integration/table_containers_test.go`
