--- examples/containers/README.md
@@ -1,3 +1,0 @@
-# Containers example
-
-Containers example demonstrates how to work with YDB container values such as `List`, `Tuple`, `Dict`, `Struct` and `Variant`
